Fix links to `modindex`
When you link to :ref:`modindex`, this links to py-modindex.html.

However our module index is `swift-modindex.html`.  This is not an issue in the AT template since we link to the swift-modindex directly, but this is an issue for linking to the index from elsewhere in the documentation (or with a different template).

The solution here is threefold:

* We duplicate the `StandardDomain` from Sphinx, calling it `SwiftStandardDomain` instead
* `SwiftStandardDomain` has `swift-` prefixes instead of `py-` prefixes
* We override from the `SwiftStandardDomain` instead of the `StandardDomain` which is the default

This is kind of terrible, but I guess nobody really planned to use Sphinx for non-Python documentation 😶

Disappointing, but there's not really a better solution.  Googling around suggested one other person has hit this, he ended up donig something pretty similar in spirit but his solution is now out of date with Sphinx changes.