Add more specific autodoc directives
This adds two new directives, `autoprotocol` and `autoextension`.  These are like `autoswift` but for (only) protocols and extensions, respectively.

This faciliates the disambiguation of tricky cases where multiple things are named the same.  See also #6 for a similar case (that cannot be disambiguated with by type alone).

Other directives should probably be added, but these are the two I need today.