Actually take config parameters
At setup time, the config parameters are not actually read.  Therefore, instead of using e.g. the `swift_search_path` that the user configured, we only use the default one.  Hope your source is in `../src`!

We now delay the indexing until the config values are actually read.  Through trial and error the `builder-inited` event seems like a good time to run the index.