Add Dockerfile
This allows docker builds of anarchysphinx.

We need various open PRs applied to build our documentation, this is the script we are using internally to build and distribute our fork.  At the moment all our patches are PRed but not all merged, and we may need other patches for future projects as necessary.

Docker images may be useful officially as well, so sending our build/packaging process upstream for inclusion.

Also, resolves #1