Add raw-members
We need a way to disambiguate between e.g.

    static func register(_: Any, name: String)

and

    static func register(legalText: String)

which share the same name.  More formally, we need a way to refer to an "exact" member.

The raw-members directive allows matching on direct sourcetext, allowing you to specify any part of a declaration to be matched, not only its name.